refactor(config): split config.go 1946→155 lines by subsystem (#135)
config.go was the single largest complexity hotspot in the backend (1946
lines). Split it into per-subsystem files while preserving 100% of the
public API (types, field names, YAML tags, Validate/ApplyDefaults behavior).

Layout after split:
- config.go (155 lines) — top-level Config struct + Load/Save/EncryptConfigFile
  + Validate/ApplyDefaults orchestrator stubs
- validate.go (564 lines) — full per-subsystem validation logic (validateConfigDetails)
- apply_defaults.go (374 lines) — full default-setting logic (applyConfigDefaults)
- config_camera.go — CameraConfig, PushTargetConfig, VideoPresetOverrides,
  HealthOverrides, XiaomiConfig
- config_merge.go — MergeConfig, RollingEnabledValue, ResolveMergeConfig
- config_timelapse.go — TimeRange, ScheduleConfig, CameraTimelapseConfig,
  parseHHMM, mergeTimeRanges, ParseMergeDuration
- config_transcoding.go — TranscodingConfig, CameraTranscodingConfig, ResolveTranscodingConfig
- config_streaming.go — HLSConfig, StreamingConfig, WebRTCConfig, ICEServerConfig, FLVConfig
- config_health.go — HealthConfig, RediscoveryConfig, AutoDiscoverConfig, HealthLayer*,
  HealthAutoRemediationConfig, ResolveHealthOverrides, ResolvedHealthOverrides
- config_ingest.go — FTPConfig, WebDAVConfig, SRTConfig, SRTStream, RTMPConfig
- config_ai.go — AIConfig
- config_server.go — ServerConfig, StorageConfig, CleanupConfig, AuthConfig,
  RateLimitConfig, MQTTConfig, ObservabilityConfig, RemoteLogConfig,
  MetricsAuthConfig, APIKeyConfig, WebSocketConfig, memAvailableMB,
  maxSegmentDurationForMem

Constraints honored:
- Zero behavior change: all type signatures, field names, YAML tags, validation
  order, and in-place mutations preserved verbatim
- Validate/ApplyDefaults bodies moved as-is into validateConfigDetails /
  applyConfigDefaults (the orchestrator stubs delegate to them)
- The 3 pre-compiled regex vars (rePlatformName/reResolution/reBitrate) moved
  to validate.go where they're used
- All existing tests pass unchanged (config_test.go 1730 lines + 4 subsystem
  test files) — they test the public API which is identical
